Add files via upload
In this project, two binary classification algorithms, viz, Decision Tree and Random Forest, are employed to build two different models, and their respective 
performances are compared using the customer dataset sourced from a telecom company in Iran. This task aims to build two different binary classification models and 
recommend a more efficient model for identifying customers who are likely to churn, that is, customers that are likely to stop patronising the company. As part of the industry standard recommendation, the Cross Industry Standard Process for Data Mining (CRISP-DM) methodology was strictly adhered to in this project.